fix: renamed BottonNavBar.tsx to BottomNavBar.tsx and updated references
